refactor: Move eulerian_paths logic from hpp to cpp.
This might improve compile performance.
