+++ eulerian_paths.cpp
@@ -1,7 +1,8 @@
-#include "geometry.hpp"
+#include "geometry_int.hpp"
 #include "bg_operators.hpp"
 #include <vector>
 #include <utility>
+#include <unordered_set>
 
 #include "segmentize.hpp"
 #include "eulerian_paths.hpp"
@@ -10,6 +11,441 @@ namespace eulerian_paths {
 
 using std::vector;
 using std::pair;
+
+template<typename p_t>
+struct identity {
+  typedef p_t type;
+};
+
+enum struct Side : bool {
+  front,
+  back,
+};
+
+static inline Side operator!(const Side& s) {
+  switch(s) {
+   case Side::front: return Side::back;
+   case Side::back: return Side::front;
+  }
+}
+
+static inline std::ostream& operator<<(std::ostream& out, const Side& s) {
+  switch(s) {
+   case Side::front:
+     out << "front";
+     break;
+   case Side::back:
+     out << "back";
+     break;
+  }
+  return out;
+}
+
+// Made public for testing.
+bool must_start_helper(size_t out_edges, size_t in_edges, size_t bidi_edges) {
+  if (out_edges > in_edges + bidi_edges) {
+    // Even with all the in and bidi paths, we would still need a path that starts here.
+    return true;
+  }
+  if (in_edges > out_edges + bidi_edges) {
+    // Pairing all bidi edges with in edges leaves no edges to start from.
+    return false;
+  }
+  // By this point, out - in <= bidi and out - in >= -bidi so abs(out-in) <=
+  // bidi If number of unmatched bidi edges is odd then this must be a start.
+  // (bidi - abs(out - in)) % 2 works but we can avoid the abs by just addings.
+  return (bidi_edges + out_edges + in_edges) % 2 == 1;
+}
+
+/* This class holds on to all the paths and uses std::multimap internally to
+ * make it quick to look up which paths extend from a given vertex and in which
+ * direction. */
+template <typename point_t, typename linestring_t>
+class path_manager {
+ public:
+  // The bool indicates if the path is reversible.
+  path_manager(const std::vector<std::pair<linestring_t, bool>>& paths) {
+    for (auto const& path : paths) {
+      add_path(path.first, path.second);
+    }
+  }
+  void add_path(const linestring_t& path, bool reversible) {
+    paths.push_back(std::make_pair(path, reversible));
+    size_t index = paths.size() - 1;
+    if (path.size() < 2) {
+      // Valid path must have a start and end.
+      return;
+    }
+    point_t start = path.front();
+    point_t end = path.back();
+    all_start_vertices.insert(start);
+    if (reversible) {
+      bidi_vertex_to_unvisited_path_index.emplace(start, std::make_pair(index, Side::front));
+      bidi_vertex_to_unvisited_path_index.emplace(end, std::make_pair(index, Side::back));
+      all_start_vertices.insert(end);
+    } else {
+      start_vertex_to_unvisited_path_index.emplace(start, std::make_pair(index, Side::front));
+      end_vertex_to_unvisited_path_index.emplace(end, std::make_pair(index, Side::back));
+    }
+  }
+  auto& get_all_start_vertices() const {
+    return all_start_vertices;
+  }
+  auto& get_start_vertex_to_unvisited_path_index() const {
+    return start_vertex_to_unvisited_path_index;
+  }
+  auto& get_bidi_vertex_to_unvisited_path_index() const {
+    return bidi_vertex_to_unvisited_path_index;
+  }
+  auto& get_end_vertex_to_unvisited_path_index() const {
+    return end_vertex_to_unvisited_path_index;
+  }
+  auto& get_path(size_t index) const {
+    return paths[index];
+  }
+  void remove_path(size_t index) {
+    auto& path = paths[index].first;
+    auto reversible = paths[index].second;
+    if (reversible) {
+      // Bidi path.
+      auto range = bidi_vertex_to_unvisited_path_index.equal_range(path.front());
+      for (auto it = range.first; it != range.second; it++) {
+        if (it->second.first == index) {
+          bidi_vertex_to_unvisited_path_index.erase(it);
+          break;
+        }
+      }
+      range = bidi_vertex_to_unvisited_path_index.equal_range(path.back());
+      for (auto it = range.first; it != range.second; it++) {
+        if (it->second.first == index) {
+          bidi_vertex_to_unvisited_path_index.erase(it);
+          break;
+        }
+      }
+    } else {
+      // Non-bidi path.
+      auto range = start_vertex_to_unvisited_path_index.equal_range(path.front());
+      for (auto it = range.first; it != range.second; it++) {
+        if (it->second.first == index) {
+          start_vertex_to_unvisited_path_index.erase(it);
+          break;
+        }
+      }
+      range = end_vertex_to_unvisited_path_index.equal_range(path.back());
+      for (auto it = range.first; it != range.second; it++) {
+        if (it->second.first == index) {
+          end_vertex_to_unvisited_path_index.erase(it);
+          break;
+        }
+      }
+    }
+  }
+ private:
+  std::vector<std::pair<linestring_t, bool>> paths;
+  // Create a map from vertex to each path that start at that vertex.
+  // It's a map to an index into the input paths.  The bool tells us
+  // if the point_t is at the front or back.  For start, it will
+  // always be true.
+  std::multimap<point_t, std::pair<size_t, Side>> start_vertex_to_unvisited_path_index;
+  // Create a map from vertex to each bidi path that may start or end
+  // at that vertex.  It's a map to an index into the input paths.
+  // The bool tells us if the point_t is at the front or back.  For
+  // bidi, it could be either.
+  std::multimap<point_t, std::pair<size_t, Side>> bidi_vertex_to_unvisited_path_index;
+  // Create a map from vertex to each path that may start or end at
+  // that vertex.  It's a map to an index into the input paths.  The
+  // bool tells us if the point_t is at the front or back.  For end,
+  // it will always be false.
+  std::multimap<point_t, std::pair<size_t, Side>> end_vertex_to_unvisited_path_index;
+  // Only the ones that have at least one potential edge leading out.
+  std::set<point_t> all_start_vertices;
+};
+
+/* This finds a minimal number of eulerian paths that cover the input.  The
+ * number of paths returned is equal to the number of vertices with odd edge
+ * count divided by 2 if all of them are bidirectional.
+ *
+ * To use, first get paths.  Each path is a vector of n points that
+ * represents n-1 line segments.  Each path is considerd
+ * bidirectional.
+ *
+ * After adding paths, build the Eulerian paths.  The resulting paths
+ * cover all segments in the input paths with the minimum number of
+ * paths as described above.
+ */
+template <typename point_t, typename linestring_t>
+class eulerian_paths {
+ public:
+  eulerian_paths(const std::vector<std::pair<linestring_t, bool>>& paths) :
+    paths(paths) {}
+  std::vector<std::pair<linestring_t, bool>> get() {
+    /* We use Hierholzer's algorithm to find the minimum cycles.  First, make a
+     * path from each vertex with more paths out than in.  In the reversible
+     * case, that means an odd path count.  Follow the path until it ends.
+     *
+     * For the reversible case, the start vertex will have count decrease by
+     * one, the end vertex will decrease from 1 to 0, and all other vertices
+     * will decrease by two.  So the result is that two vertices will have the
+     * edges counts go from odd to even and the rest will have edge counts stay
+     * either odd or even.
+     *
+     * For non-reversible, we only start at vertices where the out edges is
+     * greater than in edges.  A path must start at those because any path that
+     * goes through will subtract one from both inbound and outbound edge counts
+     * so eventually the vertex will end with only outbound eddges and so must
+     * be the start of a path.  For the new path, the start vertex will have
+     * outbound count decreased by one.  The end path will have inbound
+     * decreased by one and outbound must be zero.  All middle vertices will
+     * have inbound and outbound decreased by one each.  The result is that the
+     * outbound minus inbound count for each vertex stays the same except for
+     * the start where it goes one down by one and the end where it goes up by
+     * one.  The start was positive because we only started at edges where
+     * outbound was greater than inbound and the end must have been negative
+     * because outbound is zero.  So both the start and end vertex outbound
+     * minus inbound move closer to zero by one.  Doing this on all vertices
+     * where outbound is greater than inbound will bring all those vertices to
+     * outbound==inbound.  And because the total_outbound==total_inbound, that
+     * means that all vertices will have the same number of outbound and
+     * inbound, which means that we have made the precondition to stitch_loops.
+     */
+
+    std::vector<std::pair<linestring_t, bool>> euler_paths;
+    for (const auto& vertex : paths.get_all_start_vertices()) {
+      while (must_start(vertex)) {
+        // Make a path starting from vertex with odd count.
+        std::pair<linestring_t, bool> new_path({vertex}, true);
+        while (insert_one_path(&new_path, new_path.first.size()-1) > 0) {
+          // Keep going.
+        }
+        euler_paths.push_back(new_path);
+      }
+      // The vertex is no longer must_start.  So it must have the same or fewer
+      // out edges than in edges, even accounting for bidi edges becoming in
+      // edges.  Any path that passes into the vertex will either pass back out,
+      // removing one in edge and one out edge, or get stuck because there are
+      // zero out edges.  In either case, the number of out edges <= in edges.
+    }
+    // All vertices have out edges <= in edges.  But total out edges == total in
+    // edges so all vertices must have an equal number of out and in edges.  So
+    // if we make a path from one, it is sure to end back where it started.
+    // We'll go over all our current Euler paths and stitch in loops anywhere
+    // that there is an unvisited edge.
+    for (auto& euler_path : euler_paths) {
+      stitch_loops(&euler_path);
+    }
+
+    // Anything remaining is loops on islands.  Make all those paths, too.
+    for (const auto& vertex : paths.get_all_start_vertices()) {
+      std::pair<linestring_t, bool> new_path({vertex}, true);
+      while (insert_one_path(&new_path, new_path.first.size()-1) > 0) {
+        // Keep going.
+      }
+      if (new_path.first.size() > 1) {
+        // We can stitch right now because all vertices already have even number
+        // of edges.
+        stitch_loops(&new_path);
+        euler_paths.push_back(new_path);
+      }
+    }
+
+    return euler_paths;
+  }
+
+ private:
+  bool must_start(const point_t& vertex) const {
+    // A vertex must be a starting point if there are more out edges than in
+    // edges, even after using the bidi edges.
+    auto out_edges = paths.get_start_vertex_to_unvisited_path_index().count(vertex);
+    auto in_edges = paths.get_end_vertex_to_unvisited_path_index().count(vertex);
+    auto bidi_edges = paths.get_bidi_vertex_to_unvisited_path_index().count(vertex);
+    return must_start_helper(out_edges, in_edges, bidi_edges);
+  }
+
+  // Higher score is better.
+  template <typename p_t>
+  double path_score(const linestring_t& path_so_far,
+                    const size_t where_to_start,
+                    const std::pair<point_t, std::pair<size_t, Side>>& option,
+                    identity<p_t>) {
+    if (where_to_start < 1 || paths.get_path(option.second.first).first.size() < 2) {
+      // Doesn't matter, pick any.
+      return 0;
+    }
+    auto p0 = path_so_far[where_to_start-1];
+    auto p1 = path_so_far[where_to_start];
+    auto p2 = paths.get_path(option.second.first).first[1];
+    if (option.second.second == Side::back) {
+      // This must be reversed.
+      p2 = paths.get_path(option.second.first).first[paths.get_path(option.second.first).first.size()-2];
+    }
+
+    // cos(theta) = (a dot b)/(|a|*|b|)
+    // We don't need to take the cosine because it is decreasing over
+    // the range of theta that we care about, so they are comparable.
+    auto delta_x10 = p0.x() - p1.x();
+    auto delta_y10 = p0.y() - p1.y();
+    auto delta_x12 = p2.x() - p1.x();
+    auto delta_y12 = p2.y() - p1.y();
+    auto length_product = sqrt((delta_x10*delta_x10 + delta_y10*delta_y10) * (delta_x12*delta_x12 + delta_y12*delta_y12));
+    auto dot_product = (delta_x10*delta_x12) + (delta_y10*delta_y12);
+    return -dot_product/length_product;
+  }
+
+  double path_score(const linestring_t&,
+                    const size_t,
+                    const std::pair<point_t, std::pair<size_t, Side>>&,
+                    identity<int>) {
+    return 0;
+  }
+
+  template <typename p_t>
+  double path_score(const linestring_t& path_so_far,
+                    const size_t where_to_start,
+                    const std::pair<point_t, std::pair<size_t, Side>>& option) {
+    return path_score(path_so_far, where_to_start, option, identity<p_t>());
+  }
+
+  // Pick the best path to continue on given the path_so_far and a
+  // range of options.  The range must have at least one element in
+  // it.
+  typename std::multimap<point_t, std::pair<size_t, Side>>::const_iterator select_path(
+      const linestring_t& path_so_far,
+      const size_t where_to_start,
+      const std::pair<typename std::multimap<point_t, std::pair<size_t, Side>>::const_iterator,
+                      typename std::multimap<point_t, std::pair<size_t, Side>>::const_iterator>& options) {
+    auto best = options.first;
+    double best_score = path_score<point_t>(path_so_far, where_to_start, *best);
+    for (auto current = options.first; current != options.second; current++) {
+      double current_score = path_score<point_t>(path_so_far, where_to_start, *current);
+      if (current_score > best_score) {
+        best = current;
+        best_score = current_score;
+      }
+    }
+    return best;
+  }
+
+  // Given a point, make a path from that point as long as possible
+  // until a dead end.  Assume that point itself is already in the
+  // list.  Return the number of elements inserted.
+  size_t insert_one_path(std::pair<linestring_t, bool>* new_path, const size_t where_to_start) {
+    // Find an unvisited path that leads from point.  Prefer out edges to bidi
+    // because we may need to save the bidi edges to later be in edges.
+    auto vertex_and_path_range = paths.get_start_vertex_to_unvisited_path_index().equal_range(new_path->first.at(where_to_start));
+    if (vertex_and_path_range.first == vertex_and_path_range.second) {
+      vertex_and_path_range = paths.get_bidi_vertex_to_unvisited_path_index().equal_range(new_path->first.at(where_to_start));
+      if (vertex_and_path_range.first == vertex_and_path_range.second) {
+        // No more paths to follow.
+        return 0;
+      }
+    }
+    auto vertex_and_path_index = select_path(new_path->first, where_to_start, vertex_and_path_range);
+    size_t path_index = vertex_and_path_index->second.first;
+    Side side = vertex_and_path_index->second.second;
+    const auto& path = paths.get_path(path_index).first;
+    if (side == Side::front) {
+      // Append this path in the forward direction.
+      new_path->first.insert(new_path->first.begin() + where_to_start + 1, path.cbegin()+1, path.cend());
+    } else {
+      // Append this path in the reverse direction.
+      new_path->first.insert(new_path->first.begin() + where_to_start + 1, path.crbegin()+1, path.crend());
+    }
+    paths.remove_path(path_index);
+    new_path->second = new_path->second && paths.get_path(path_index).second;
+    return path.size() - 1;
+  }
+
+  // Only call this when there are no vertices with uneven edge count.  That
+  // means that all vertices must have as many edges leading in as edges leading
+  // out.  This can be true if a vertex has no paths at all.  This is also true
+  // if some edges are reversable and they could poentially be used to make the
+  // number of in edges equal to the number of out edges.  This will traverse a
+  // path and, if it finds an unvisited edge, will make a Euler circuit there
+  // and stitch it into the current path.  Because all paths have the same
+  // number of in and out, the stitch can only possibly end in a loop.  This
+  // continues until the end of the path.
+  void stitch_loops(std::pair<linestring_t, bool> *euler_path) {
+    // Use a counter and not a pointer because the list will grow and pointers
+    // may be invalidated.
+    for (size_t i = 0; i < euler_path->first.size(); i++) {
+      // Make a path from here.  We don't need the first element, it's already in our path.
+      auto j = i;
+      while (true) {
+        auto elements_inserted = insert_one_path(euler_path, j);
+        if (elements_inserted == 0) {
+          break;
+        }
+        j += elements_inserted;
+      }
+    }
+  }
+  path_manager<point_t, linestring_t> paths;
+}; //class eulerian_paths
+template <typename point_t, typename linestring_t>
+bool check_eulerian_paths(const std::vector<std::pair<linestring_t, bool>>& before,
+                          const std::vector<std::pair<linestring_t, bool>>& after) {
+  std::unordered_multiset<std::pair<std::pair<point_t, point_t>, bool>> all_edges;
+  for (auto const& linestring : before) {
+    auto const reversible = linestring.second;
+    auto const path = linestring.first;
+    if (path.size() < 2) {
+      continue;
+    }
+    for (size_t i = 0; i < path.size()-1; i++) {
+      auto const p0 = path[i];
+      auto const p1 = path[i+1];
+      auto const edge = std::make_pair(p0, p1);
+      all_edges.insert({edge, reversible});
+    }
+  }
+  for (auto const& linestring : after) {
+    auto const path = linestring.first;
+    for (size_t i = 0; i < path.size()-1; i++) {
+      auto const p0 = path[i];
+      auto const p1 = path[i+1];
+      auto const edge = std::make_pair(p0, p1);
+      auto const reversed_edge = std::make_pair(p1, p0);
+      // If the edge is in all_edges twice, reversible and not, we prefer to delete the non-reversible one first because the other one can be more
+      // flexible for searching later.
+      if (all_edges.find({edge, false}) != all_edges.end()) {
+        all_edges.erase(all_edges.find({edge, false}));
+      } else if (all_edges.find({edge, true}) != all_edges.end()) {
+        all_edges.erase(all_edges.find({edge, true}));
+      } else if (all_edges.find({reversed_edge, true}) != all_edges.end()) {
+        all_edges.erase(all_edges.find({reversed_edge, true}));
+      } else {
+        return false;
+      }
+    }
+  }
+  if (all_edges.size() > 0) {
+    return false;
+  }
+  return true;
+}
+
+template
+bool check_eulerian_paths<point_type>(const std::vector<std::pair<linestring_type, bool>>& before,
+                                      const std::vector<std::pair<linestring_type, bool>>& after);
+template
+bool check_eulerian_paths<int>(const std::vector<std::pair<std::vector<int>, bool>>& before,
+                               const std::vector<std::pair<std::vector<int>, bool>>& after);
+
+// Returns a minimal number of toolpaths that include all the milling in the
+// oroginal toolpaths.  Each path is traversed once.  Each path has a bool
+// indicating if the path is reversible.
+template <typename point_t, typename linestring_t>
+std::vector<std::pair<linestring_t, bool>> get_eulerian_paths(const std::vector<std::pair<linestring_t, bool>>& paths) {
+  auto const ret = eulerian_paths<point_t, linestring_t>(
+      paths).get();
+  assert((check_eulerian_paths<point_t, linestring_t>(paths, ret)));
+  return ret;
+}
+
+template
+std::vector<std::pair<std::vector<int>, bool>> get_eulerian_paths<int>(const std::vector<std::pair<std::vector<int>, bool>>& paths);
+template
+std::vector<std::pair<linestring_type, bool>> get_eulerian_paths<point_type>(const std::vector<std::pair<linestring_type, bool>>& paths);
 
 // This calls segmentize and then get_eulerian_paths.  If unique is
 // true, remove repeated segments.
